backend/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
faiss-cpu
youtube-transcript-api
python-dotenv
tiktoken